telegram-voice-transcriber/requirements.txt
fastapi
uvicorn[standard]
openai
httpx
python-dotenv
pydub
